wip: Split src/lib.rs into submodules
