+++ src/configuration.rs
@@ -1,0 +1,1 @@
+
