+++ src/lib.rs
@@ -1,28 +1,3 @@
-use actix_web::dev::Server;
-use actix_web::{web, App, HttpRequest, HttpResponse, HttpServer, Responder};
-use std::net::TcpListener;
-
-async fn health_check() -> impl Responder {
-    HttpResponse::Ok()
-}
-
-#[derive(serde::Deserialize)]
-struct FormData {
-    email: String,
-    name: String,
-}
-
-async fn subscribe(_form: web::Form<FormData>) -> impl Responder {
-    HttpResponse::Ok()
-}
-
-pub fn run(listener: TcpListener) -> std::io::Result<Server> {
-    let server = HttpServer::new(|| {
-        App::new()
-            .route("/health_check", web::get().to(health_check))
-            .route("/subscriptions", web::post().to(subscribe))
-    })
-    .listen(listener)?
-    .run();
-    Ok(server)
-}
+pub mod configuration;
+pub mod routes;
+pub mod startup;
