+++ src/routes/health_check.rs
@@ -1,0 +1,5 @@
+use actix_web::{HttpResponse, Responder};
+
+pub async fn health_check() -> impl Responder {
+    HttpResponse::Ok()
+}
