+++ src/routes/mod.rs
@@ -1,0 +1,5 @@
+pub mod health_check;
+pub mod subscriptions;
+
+pub use health_check::*;
+pub use subscriptions::*;
