+++ src/startup.rs
@@ -1,0 +1,15 @@
+use crate::routes::{health_check, subscribe};
+use actix_web::dev::Server;
+use actix_web::{web, App, HttpServer};
+use std::net::TcpListener;
+
+pub fn run(listener: TcpListener) -> std::io::Result<Server> {
+    let server = HttpServer::new(|| {
+        App::new()
+            .route("/health_check", web::get().to(health_check))
+            .route("/subscriptions", web::post().to(subscribe))
+    })
+    .listen(listener)?
+    .run();
+    Ok(server)
+}
