+++ tests/health_check.rs
@@ -3,7 +3,7 @@ use std::net::TcpListener;
 fn spawn_app() -> String {
     let listener = TcpListener::bind("127.0.0.1:0").expect("Failed to bind random port");
     let port = listener.local_addr().unwrap().port();
-    let server = newsletter_lib::run(listener).expect("Failed to bind address");
+    let server = newsletter_lib::startup::run(listener).expect("Failed to bind address");
     tokio::spawn(server);
     format!("http://127.0.0.1:{}", port)
 }
